break up fuel type options into own files
