+++ src/app/facility/utility-data/energy-consumption/energy-source/edit-meter-form/emissions-details-table/emissions-details-table.component.ts
@@ -1,6 +1,6 @@
 import { Component, Input } from '@angular/core';
-import { FuelTypeOption } from '../editMeterOptions';
 import { FormGroup } from '@angular/forms';
+import { FuelTypeOption } from 'src/app/shared/fuel-options/fuelTypeOption';
 
 @Component({
   selector: 'app-emissions-details-table',
